More dynamic way to find the config
This Commit is inspired and uses code from dehydrated. Th config will now
be searched inside in the normal posix lokations first and than in the current dir.
The last possible folder is the dir the script iss living in.

This should be more portable than the hardcoded path to /opt/hpkpinx/ in the
script until now, expecially as the git is named HPKPinx and just cloning
it to /opt results in an not found config.