Allow multible certificates to be handeld
The first new option MULTIBLE_HPKP_CONF tells the script to generate the nginx
config file name from the common name of the certificate currently
handled. This allow the script to handle more than one Domain with different
private keys.

The second new config variable STATIC_PIN_FILE  allows it to define a file
from which the STATIC_PIN for each common name is read. This basically allows
it to have an different backup key per CN. The file is organized line wise
separated by a space the CN and the HPKP static pin.